allow fractional components in ratio specifications

+++ src/Rules.java
@@ -262,8 +262,8 @@ public class Rules
             final double Result;
             if (SepPos >= 0)
               {
-                final double Numer = Integer.parseInt(s.substring(0, SepPos));
-                final double Denom = Integer.parseInt(s.substring(SepPos + 1, s.length()));
+                final double Numer = Double.parseDouble(s.substring(0, SepPos));
+                final double Denom = Double.parseDouble(s.substring(SepPos + 1, s.length()));
                 if (Denom <= 0.0 || Numer <= 0.0)
                   {
                     throw new NumberFormatException("ratio cannot be zero or negative");
